Refactor context manager package
